move listener to condition
